Fixed modal state to read areFilesBeingCreated and files that got removed at wrong index

+++ Attest/View/FilesView.swift
@@ -13,12 +13,13 @@ struct FilesView: View {
     @EnvironmentObject var device : Device
     @State private var displayModal: Bool = false
 
+    
     var body: some View {
         NavigationView {
             VStack{
                 if store.files.count > 0 {
-                    if device.capacity != nil {
-                        Text("Free disk space: \(device.capacity!.sizeString())")
+                    if device.freeCapacity != nil {
+                        Text("Free disk space: \(device.freeCapacity!.sizeString())")
                             .padding([.top, .bottom], 10)
                     }
                     List {
@@ -33,8 +34,8 @@ struct FilesView: View {
                     VStack {
                         Text("No files yet")
                             .font(.largeTitle)
-                        if device.capacity != nil {
-                            Text("Create files to see a list of your Documents folder here. Free disk space: \(device.capacity!.sizeString())")
+                        if device.freeCapacity != nil {
+                            Text("Create files to see a list of your Documents folder here. Free disk space: \(device.freeCapacity!.sizeString())")
                                 .font(.body)
                                 .lineLimit(3)
                         } else {
@@ -72,30 +73,40 @@ struct FilesView: View {
 
     
     private func delete(at offsets: IndexSet) {
-        let fileToBeRemoved : File = store.files.removeLast()
-        store.removeFile(file: fileToBeRemoved)
-        self.device.readDeviceProperties()
+        if offsets.count > 1 {
+            print("Got offsets for list deletion > 1")
+            return
+        }
+        
+        if let index = offsets.first {
+            let fileToBeRemoved : File = store.files.remove(at: index)
+            store.removeFile(file: fileToBeRemoved)
+            self.device.readDeviceProperties()
+        }
     }
     
 }
 
 struct FilesViewView_Previews: PreviewProvider {
     static var previews: some View {
+        
         Group {
             FilesView()
                 .environmentObject(Device())
                 .environmentObject(FileStore(files: []))
+                .previewDevice(PreviewDevice(rawValue: "iPhone SE"))
             
             FilesView()
                 .environmentObject(Device())
                 .environmentObject(FileStore(files: testData))
+                .previewDevice(PreviewDevice(rawValue: "iPhone SE"))
             
             FilesView()
                 .environmentObject(Device())
                 .environmentObject(FileStore(files: testData))
                 .environment(\.colorScheme, .dark)
+            .previewDevice(PreviewDevice(rawValue: "iPhone XS Max"))
         }
-
     }
 }
